Services/requirements.txt
grpcio
grpcio-tools

requests